refactor walletTable
