+++ src/client/wallet/WalletTable/WalletTable-nightmode.less
@@ -1,4 +1,4 @@
-@import (reference) '../styles/custom.less';
+@import (reference) '../../styles/custom.less';
 
 .nightmode {
   .WalletTable {
