+++ src/client/wallet/WalletTable/WalletTable.less
@@ -1,6 +1,6 @@
-@import (reference) '../styles/modules/variables.less';
-@import (reference) '../styles/custom.less';
-@import (reference) '../styles/mixins.less';
+@import (reference) '../../styles/modules/variables.less';
+@import (reference) '../../styles/custom.less';
+@import (reference) '../../styles/mixins.less';
 
 .WalletTable {
   border: 1px solid @black;
@@ -141,4 +141,4 @@
   margin-bottom: -30px;
 }
 
-@import 'WalletTable-nightmode';
+@import './WalletTable-nightmode';
