+++ src/client/wallet/constants.js
@@ -1,0 +1,34 @@
+const WalletTableHeader = [
+  {
+    className: 'WalletTable__date',
+    id: 'table_date',
+    message: 'Date',
+  },
+  {
+    className: 'WalletTable__HIVE',
+    id: 'table_HIVE',
+    message: 'HIVE',
+  },
+  {
+    className: 'WalletTable__HP',
+    id: 'table_HP',
+    message: 'HP',
+  },
+  {
+    className: 'WalletTable__HBD',
+    id: 'table_HBD',
+    message: 'HBD',
+  },
+  {
+    className: 'WalletTable__description',
+    id: 'table_description',
+    message: 'Description',
+  },
+  {
+    className: 'WalletTable__memo',
+    id: 'memo',
+    message: 'Memo',
+  },
+];
+
+export default WalletTableHeader;
